Move system store management from snapstate to storestate for consistency.
