+++ overlord/snapstate/storehelpers.go
@@ -22,6 +22,7 @@ package snapstate
 import (
 	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/state"
+	"github.com/snapcore/snapd/overlord/storestate"
 	"github.com/snapcore/snapd/snap"
 	"github.com/snapcore/snapd/store"
 )
@@ -55,7 +56,7 @@ func updateInfo(st *state.State, snapst *SnapState, channel string, userID int) 
 		Epoch:    curInfo.Epoch,
 	}
 
-	theStore := Store(st)
+	theStore := storestate.Store(st)
 	st.Unlock() // calls to the store should be done without holding the state lock
 	res, err := theStore.LookupRefresh(refreshCand, user)
 	st.Lock()
@@ -67,7 +68,7 @@ func snapInfo(st *state.State, name, channel string, revision snap.Revision, use
 	if err != nil {
 		return nil, err
 	}
-	theStore := Store(st)
+	theStore := storestate.Store(st)
 	st.Unlock() // calls to the store should be done without holding the state lock
 	spec := store.SnapSpec{
 		Name:     name,
